==> dimensional/02_create_facts.ipynb <==
import pandas as pd
import numpy as np
import random

## Leer datos

dim_hour = pd.read_csv('git/shorts/dimensional/data/dim_hour.csv')
print(dim_hour.shape)
display(dim_hour.head(3))

dim_date = pd.read_csv('git/shorts/dimensional/data/dim_date.csv')
print(dim_date.shape)
display(dim_date.head(3))

dim_date_open_store = pd.read_csv('git/shorts/dimensional/data/dim_date.csv')
print(dim_date_open_store.shape)
display(dim_date_open_store.head(3))

dim_product = pd.read_csv('git/shorts/dimensional/data/dim_product.csv')
print(dim_product.shape)
display(dim_product.head(3))

dim_store = pd.read_csv('git/shorts/dimensional/data/dim_store.csv')
print(dim_store.shape)
display(dim_store.head(3))

dim_promotion = pd.read_csv('git/shorts/dimensional/data/dim_promotion.csv')
print(dim_promotion.shape)
print(dim_promotion.head(6))

## Retail - Transacciones

np.random.seed(15)
random.seed(15)

transactions = pd.DataFrame({'id_transaction': np.arange(1, 10001)})
transactions['id_date'] = np.random.choice(dim_date_open_store['id_date'], size=10000)
transactions['id_hour'] = random.choices([11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21],
                                         weights=[0.05, 0.05, 0.05, 0.10, 0.10,
                                                  0.10, 0.15, 0.15, 0.10, 0.10,
                                                  0.05], k=10000)
transactions['id_product'] = np.random.choice(dim_product['id_product'], size=10000)
transactions['id_store'] = random.choices([1, 2, 3, 4, 5, 6, 7, 8, 9, 10,
                                           11, 12, 13, 14, 15, 16, 17, 18, 19, 20],
                                           weights=[0.03, 0.05, 0.07, 0.05, 0.05,
                                                    0.04, 0.06, 0.01, 0.09, 0.05,
                                                    0.06, 0.04, 0.05, 0.08, 0.02,
                                                    0.05, 0.09, 0.01, 0.03, 0.07], k=10000)

transactions['pos_transaction'] = np.arange(1, 10001)
transactions['id_promotion'] = random.choices([1, 2, 3, 4, 5, 6], weights=[0.90, 0.02, 0.03, 0.01, 0.02, 0.02], k=10000)
transactions['sales_quantity'] = random.choices([1, 2, 3], weights=[0.92, 0.07, 0.01], k=10000)
transactions['cost_factor'] = np.round(np.random.normal(loc=0.60, scale=0.10, size=10000), 2)

transactions.loc[transactions['id_promotion'].isin([2, 3, 4]), 'sales_quantity'] = 2
transactions['regular_unit_price'] = np.round(np.random.normal(loc=500.00, scale=150.00, size=10000), 0)
transactions['cost_unit_price'] = transactions['regular_unit_price'] * transactions['cost_factor']
transactions = transactions.drop(columns=['cost_factor'])

transactions = transactions.loc[transactions['id_date'] >= 20150101]
transactions = transactions.loc[transactions['id_date'] <= 20301231]
transactions = transactions.loc[transactions['regular_unit_price'] > 100]

transactions['discount_unit_price'] = 0.0
index = transactions.loc[transactions['id_promotion'].isin([2, 3, 4])].index
transactions.loc[index, 'discount_unit_price'] = transactions.loc[index, 'regular_unit_price'] * 0.5
index = transactions.loc[transactions['id_promotion'].isin([5, 6])].index 
transactions.loc[index, 'discount_unit_price'] = 50.0

transactions['net_unit_price'] = transactions['regular_unit_price'] - transactions['discount_unit_price']

transactions['extended_discount_price'] = transactions['discount_unit_price'] * transactions['sales_quantity']
transactions['extended_net_price'] = transactions['net_unit_price'] * transactions['sales_quantity']
transactions['extended_cost_price'] = transactions['cost_unit_price'] * transactions['sales_quantity']
transactions['extended_gross_profit'] = transactions['extended_net_price'] - transactions['extended_cost_price']

transactions = transactions.merge(dim_store[['id_store', 'open_date']], on='id_store', how='inner')
transactions = transactions.loc[transactions['id_date'] >= transactions['open_date']]
transactions = transactions.drop(columns=['open_date'])

print(transactions.shape)
display(transactions.head(10))

transactions.to_csv('git/shorts/dimensional/data/fact_retail_transactions.csv', index=False)

#COBERTURA DE PRODUCTOS NO VENDIDOS
sub_a = dim_store[['id_store']].drop_duplicates()
sub_a['cross'] = True

sub_b = dim_product[['genre']].drop_duplicates()
sub_b['cross'] = True

sub_c = dim_date.loc[dim_date['month'] >= 7]
sub_c = sub_c.loc[sub_c['year'] == 2025]
sub_c = sub_c[['month']].drop_duplicates()
sub_c['cross'] = True

cross_product = sub_a.merge(sub_b, on='cross', how='inner')
cross_product = cross_product.merge(sub_c, on='cross', how='inner')
cross_product = cross_product.drop(columns=['cross'])
cross_product['incidence_count'] = 1

display(cross_product.head(10))

cross_product.to_csv('git/shorts/dimensional/data/fact_retail_coverage_genre.csv', index=False)

## Inventario - Periodico

sub_a = dim_store[['id_store']].drop_duplicates()
sub_a['cross'] = True

sub_b = dim_product[['id_product']].drop_duplicates()
sub_b['cross'] = True

cross_product = sub_a.merge(sub_b, on='cross', how='inner')
cross_product = cross_product.drop(columns=['cross'])
cross_product['initial_stock'] = 100

display(cross_product.head(10))


ventas = transactions.groupby(['id_store', 'id_product', 'id_date']).agg({'sales_quantity': 'count'}).reset_index()
print(ventas.shape)
display(ventas.head(10))

lista_fechas = dim_date.loc[dim_date['month'] >= 3]
lista_fechas = lista_fechas['id_date'].to_list()
print(lista_fechas[:10])

ventas_periodicas = []
for fecha in lista_fechas:
    temp = ventas.loc[(ventas['id_date'] <= fecha) & (ventas['id_date'] >= 20250301)]
    temp = temp.groupby(['id_store', 'id_product']).agg({'sales_quantity': 'sum'}).reset_index()
    
    temp_2 = ventas.loc[ventas['id_date'] == fecha]
    temp_2 = temp_2[['id_store', 'id_product', 'sales_quantity']]
    temp_2 = temp_2.rename(columns={'sales_quantity': 'quantity_sold'})

    temp = cross_product.merge(temp, on=['id_store', 'id_product'], how='left')
    temp = temp.merge(temp_2, on=['id_store', 'id_product'], how='left')
    temp['sales_quantity'] = temp['sales_quantity'].fillna(0).astype(int)
    temp['quantity_sold'] = temp['quantity_sold'].fillna(0).astype(int)
    temp['id_date'] = fecha
    temp['quantity_on_hand'] = temp['initial_stock'] - temp['sales_quantity']

    ventas_periodicas.append(temp)

ventas_periodicas = pd.concat(ventas_periodicas)
ventas_periodicas = ventas_periodicas[['id_store', 'id_product', 'id_date', 'quantity_on_hand', 'quantity_sold']]
print(ventas_periodicas.shape)
display(ventas_periodicas.head(10))

ventas_periodicas.to_csv('git/shorts/dimensional/data/fact_inventory_periodic.csv', index=False)

## Inventario - Transacciones

random.seed(15)

inventory_transactions = pd.DataFrame({'id_inventory_transaction': np.arange(1, 21)})
inventory_transactions['id_store'] = [1] * 10 + [2] * 10
inventory_transactions['id_product'] = [1, 1, 1, 1, 1, 2, 2, 2, 2, 2] * 2
inventory_transactions['id_date'] = [20250301, 20250305, 20250310, 20250315, 20250320,
                                     20250301, 20250305, 20250310, 20250315, 20250320] * 2
inventory_transactions['id_inventory_transaction_type'] = [1, 1, 2, 3, 6, 1, 1, 2, 3, 6,
                                                           1, 1, 2, 3, 4, 1, 1, 2, 3, 5]
inventory_transactions['quantity'] = random.choices([10, 20, 30, 40, 50], k=20)
inventory_transactions['amount'] = inventory_transactions['quantity'] * 500.0


display(inventory_transactions.head(3))

inventory_transactions.to_csv('git/shorts/dimensional/data/fact_inventory_transactions.csv', index=False)

## Inventario - Acumulable

inventory_accumulated = pd.DataFrame({'id_inventory_accumulated': np.arange(1, 4)})
inventory_accumulated['product_lot'] = ['L001', 'L002', 'L003']
inventory_accumulated['id_store'] = [1, 1, 1]
inventory_accumulated['id_product'] = [1, 2, 3]
inventory_accumulated['id_date_received'] = [20250301, 20250301, 20250301]
inventory_accumulated['id_date_inspected'] = [20250303, 20250303, 20250304]
inventory_accumulated['id_date_bin_placement'] = [21000101, 20250310, 20250306]
inventory_accumulated['id_date_initial_shipment'] = [21000101, 20250315, 20250306]
inventory_accumulated['id_date_last_shipment'] = [21000101, 21000101, 20250401]
inventory_accumulated['quantity_received'] = [100, 100, 50]
inventory_accumulated['quantity_inspected'] = [100, 100, 50]
inventory_accumulated['quantity_returned_to_vendor'] = [0, 1, 0]
inventory_accumulated['quantity_placed_in_bin'] = [0, 30, 50]
inventory_accumulated['quantity_shipped_to_customer'] = [0, 2, 49]
inventory_accumulated['quantity_returned_by_customer'] = [0, 0, 0]
inventory_accumulated['quantity_damaged'] = [0, 0, 1]

print(inventory_accumulated.shape)
display(inventory_accumulated.head(3))

inventory_accumulated.to_csv('git/shorts/dimensional/data/fact_inventory_accumulated.csv', index=False)

## Procuracion - Transacciones

procurement_transactions = pd.DataFrame({'id_procurement_transaction': np.arange(1, 5)})
procurement_transactions['produrement_transaction_date'] = [20250301, 20250305, 20250310, 20250315]
procurement_transactions['id_product'] = [1, 1, 2, 2]
procurement_transactions['id_vendor'] = [1, 2, 1, 2]
procurement_transactions['id_contract_terms'] = [1, 1, 2, 1]
procurement_transactions['contract_number'] = ['L001', 'L002', 'L001', 'L002']
procurement_transactions['quantity_ordered'] = [50, 50, 100, 100]
procurement_transactions['amount'] = [25000.0, 25000.0, 50000.0, 50000.0]

print(procurement_transactions.shape)
display(procurement_transactions.head(5))

procurement_transactions.to_csv('git/shorts/dimensional/data/fact_procurement_transactions.csv', index=False)

## Ordenes - Conversión de divisas

currency_rates = pd.DataFrame({'id_currency_rate': np.arange(1, 7)})
currency_rates['currency_rate_date'] = [20250301, 20250301, 20250301, 20250302, 20250302, 20250302]
currency_rates['id_currency_source'] = [2, 3, 4, 2, 3, 6]
currency_rates['id_currency_destination'] = [1, 1, 1, None, 1, 1]
currency_rates['source_to_destination_rate'] = [18.45, 20.01, 24.15, 18.50, 20.05, 24.20]
currency_rates['destination_to_source_rate'] = 1 / currency_rates['source_to_destination_rate']

print(currency_rates.shape)
display(currency_rates.head(10))

currency_rates.to_csv('git/shorts/dimensional/data/fact_currency_rates.csv', index=False)
