==> dimensional/03_retail_analysis.ipynb <==
import pandas as pd
import numpy as np

## Leer datos

dim_hour = pd.read_csv('git/shorts/dimensional/data/dim_hour.csv')
print(dim_hour.shape)
display(dim_hour.head(3))

dim_date = pd.read_csv('git/shorts/dimensional/data/dim_date.csv')
print(dim_date.shape)
display(dim_date.head(3))

dim_product = pd.read_csv('git/shorts/dimensional/data/dim_product.csv')
print(dim_product.shape)
display(dim_product.head(3))

dim_store = pd.read_csv('git/shorts/dimensional/data/dim_store.csv')
print(dim_store.shape)
display(dim_store.head(3))

dim_promotion = pd.read_csv('git/shorts/dimensional/data/dim_promotion.csv')
print(dim_promotion.shape)
print(dim_promotion.head(6))

transactions = pd.read_csv('git/shorts/dimensional/data/fact_retail_transactions.csv')
print(transactions.shape)
display(transactions.head(3))

factless_coverage_genre = pd.read_csv('git/shorts/dimensional/data/fact_retail_coverage_genre.csv')
print(factless_coverage_genre.shape)
display(factless_coverage_genre.head(3))

sales_by_date_and_promotion = transactions.groupby(by=['id_date', 'id_promotion'], as_index=False).agg({'extended_net_price': 'sum', 'extended_gross_profit': 'sum',})
sales_by_date_and_promotion = sales_by_date_and_promotion.merge(dim_promotion, how='inner', on='id_promotion')
sales_by_date_and_promotion = sales_by_date_and_promotion.merge(dim_date[['id_date', 'date', 'week_of_year', 'quarter_year', 'holiday_indicator']], how='inner', on='id_date')                                                                                                        

print(sales_by_date_and_promotion.shape)
display(sales_by_date_and_promotion.head(3))

sales_by_category = transactions.groupby(by=['id_product'], as_index=False).agg({'extended_net_price': 'sum'})
sales_by_category = sales_by_category.merge(dim_product[['id_product', 'genre', 'subgenre']], how='inner', on='id_product')
sales_by_category = sales_by_category.groupby(by=['genre', 'subgenre'], as_index=False).agg({'extended_net_price': 'sum'})

print(sales_by_category.shape)
display(sales_by_category.head(30))


#FACTLESS FOR GENRE COVERAGE
genre_sales = transactions[['id_product', 'id_date', 'id_store', 'sales_quantity']]
genre_sales = genre_sales.merge(dim_product[['id_product', 'genre']], how='inner', on='id_product')
genre_sales = genre_sales.merge(dim_date[['id_date', 'month']], how='inner', on='id_date')

genre_sales = genre_sales.groupby(by=['genre', 'id_store', 'month'], as_index=False).agg({'sales_quantity': 'sum'})
genre_sales = genre_sales.loc[genre_sales['month'] >= 7]

print(genre_sales.shape)
display(genre_sales.head(3))

#UNSOLD GENRES
unsold_genres = factless_coverage_genre.merge(genre_sales, how='left', on=['genre', 'id_store', 'month'])
unsold_genres = unsold_genres.loc[unsold_genres['sales_quantity'].isna()]
unsold_genres = unsold_genres.merge(dim_store[['id_store', 'store_name', 'state']], how='inner', on='id_store')
unsold_genres = unsold_genres.merge(dim_date[['month', 'month_name', 'year']].drop_duplicates(), how='inner', on='month')
unsold_genres = unsold_genres.sort_values(by=['year', 'month', 'state', 'store_name', 'genre'])
unsold_genres = unsold_genres[['year', 'month', 'month_name', 'state', 'store_name', 'genre', 'incidence_count']]

print(unsold_genres.shape)
display(unsold_genres.head(30))

#PERIODIC INVENTORY
periodic_inventory = pd.read_csv('git/shorts/dimensional/data/fact_inventory_periodic.csv')

print(periodic_inventory['id_store'].unique().shape)
print(periodic_inventory['id_product'].unique().shape)
print(periodic_inventory['id_date'].unique().shape)

print(periodic_inventory.shape)
display(periodic_inventory.head(3))

print(periodic_inventory['quantity_on_hand'].value_counts())

display(periodic_inventory.loc[periodic_inventory['quantity_sold'] > 0].sample(10))

#HECHO SEMI ADITIVO
avg_inventory = periodic_inventory.groupby(by=['id_product', 'id_store'], as_index=False).agg({'quantity_on_hand': 'mean'})
avg_inventory = avg_inventory.groupby(by=['id_product'], as_index=False).agg({'quantity_on_hand': 'mean'})

display(avg_inventory.head(30))
